Refactor Hasher to move it from :server subproject to :common, so it can be re-utilized by simulator as well.
