+++ common/build.gradle.kts
@@ -1,5 +1,8 @@
 // SPDX-License-Identifier: Apache-2.0
-plugins { id("org.hiero.gradle.module.library") }
+plugins {
+    id("org.hiero.gradle.module.library")
+    id("application")
+}
 
 description = "Commons module with logic that could be abstracted and reused."
