+++ server/src/main/java/com/hedera/block/server/verification/session/BlockVerificationSessionSync.java
@@ -1,10 +1,10 @@
 // SPDX-License-Identifier: Apache-2.0
 package com.hedera.block.server.verification.session;
 
-import static com.hedera.block.server.verification.hasher.HashingUtilities.getBlockItemHash;
+import static com.hedera.block.common.hasher.HashingUtilities.getBlockItemHash;
 
+import com.hedera.block.common.hasher.NaiveStreamingTreeHasher;
 import com.hedera.block.server.metrics.MetricsService;
-import com.hedera.block.server.verification.hasher.NaiveStreamingTreeHasher;
 import com.hedera.block.server.verification.signature.SignatureVerifier;
 import com.hedera.hapi.block.BlockItemUnparsed;
 import com.hedera.hapi.block.stream.BlockProof;
