+++ server/src/main/java/com/hedera/block/server/verification/signature/SignatureVerifierDummy.java
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: Apache-2.0
 package com.hedera.block.server.verification.signature;
 
-import com.hedera.block.server.verification.hasher.HashingUtilities;
+import com.hedera.block.common.hasher.HashingUtilities;
 import com.hedera.pbj.runtime.io.buffer.Bytes;
 import edu.umd.cs.findbugs.annotations.NonNull;
 import javax.inject.Inject;
