+++ server/src/main/java/module-info.java
@@ -25,7 +25,6 @@ module com.hedera.block.server {
     exports com.hedera.block.server.pbj;
     exports com.hedera.block.server.producer;
     exports com.hedera.block.server.verification;
-    exports com.hedera.block.server.verification.hasher;
     exports com.hedera.block.server.verification.session;
     exports com.hedera.block.server.verification.signature;
     exports com.hedera.block.server.verification.service;
