+++ server/src/test/java/com/hedera/block/server/verification/signature/SignatureVerifierDummyTest.java
@@ -3,7 +3,7 @@ package com.hedera.block.server.verification.signature;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-import com.hedera.block.server.verification.hasher.HashingUtilities;
+import com.hedera.block.common.hasher.HashingUtilities;
 import com.hedera.pbj.runtime.io.buffer.Bytes;
 import org.junit.jupiter.api.Test;
